Add corpus module
Currently the corpus has two InMemoryCorpus for VMInput and State.
Maybe later we want to change the vminput to Txn, depending on how
we define the inputs. Also we may not need that many generics, e.g.
we can define InMemoryItyCorpus as a non-generic struct and then
implement Corpus<CorpusInput> for InMemoryItyCorpus.